Update issues form code review
