+++ tests/conftest.py
@@ -5,9 +5,7 @@ To define the location of module containing fixtures, the absolute path to that 
 starting from the main package folder must be given.
 """
 
-import shutil
 from collections.abc import Generator
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, cast
 
@@ -16,6 +14,8 @@ from _pytest._code.code import ExceptionRepr
 from _pytest.nodes import Item
 from _pytest.reports import TestReport
 from _pytest.runner import CallInfo
+
+from tests.helpers import JsonFilesExporter, OutFileExporter
 
 JSON_DIR = Path(__file__).parent / "json_files"
 
@@ -153,69 +153,6 @@ def pytest_addoption(parser: pytest.Parser) -> None:
     )
 
 
-@dataclass(frozen=True)
-class JsonFilesExporter:
-    json_files_dir: Path
-    prefix: str
-    enabled: bool
-
-    def export_jsons_from(
-        self,
-        src_dir: Path,
-        *,
-        recursive: bool = False,
-        overwrite: bool = True,
-        gbw_subdir: str = "gbw",
-        property_subdir: str = "property",
-    ) -> tuple[list[Path], list[Path]]:
-        """
-        Copy JSON files from src_dir into tests/json_files/<plain_subdir|property_subdir>.
-
-        - plain:     *.json excluding *.property.json
-        - property:  *.property.json
-
-        Destination filenames are prefixed:
-            <prefix>__<original_filename>
-
-        Returns: (plain_dests, property_dests)
-        """
-        pattern = "**/*.json" if recursive else "*.json"
-        sources = sorted(p for p in src_dir.glob(pattern) if p.is_file())
-
-        plain_out = self.json_files_dir / gbw_subdir
-        prop_out = self.json_files_dir / property_subdir
-        plain_out.mkdir(parents=True, exist_ok=True)
-        prop_out.mkdir(parents=True, exist_ok=True)
-
-        plain_dests: list[Path] = []
-        prop_dests: list[Path] = []
-
-        for src in sources:
-            name = src.name
-
-            is_property = name.endswith(".property.json")
-            if is_property:
-                dst = prop_out / f"{self.prefix}_{name}"
-                prop_dests.append(dst)
-            else:
-                # it's a .json from the glob, but not a .property.json
-                dst = plain_out / f"{self.prefix}_{name}"
-                plain_dests.append(dst)
-
-            if not self.enabled:
-                continue
-
-            if dst.exists() and not overwrite:
-                raise FileExistsError(f"JSON file exists (overwrite disabled): {dst}")
-
-            shutil.copy2(src, dst)
-
-        if self.enabled and not (plain_dests or prop_dests):
-            raise FileNotFoundError(f"No JSON files found in {src_dir} (pattern={pattern!r})")
-
-        return plain_dests, prop_dests
-
-
 @pytest.fixture(scope="session")
 def json_files_dir(request: pytest.FixtureRequest) -> Path:
     return Path(request.config.rootpath) / "tests" / "json_files"
@@ -231,124 +168,6 @@ def json_files_exporter(request: pytest.FixtureRequest, json_files_dir: Path) ->
         prefix=request.node.name,  # json_file basename
         enabled=request.config.getoption("--update-json-files"),
     )
-
-
-def _is_separator(line: str) -> bool:
-    """Return True if the line contains only dashes, equals signs, and spaces (ORCA separator line)."""
-    stripped = line.strip()
-    return bool(stripped) and all(c in "-= " for c in stripped)
-
-
-def _extract_orca_blocks(src: Path) -> str:
-    """
-    Extract only the grepper-relevant blocks from a full ORCA .out file.
-
-    Strips all system-specific info (paths, hostnames, timings) and keeps only:
-      - Total Charge / Multiplicity lines (once each)
-      - Each CARTESIAN COORDINATES (ANGSTROEM) block
-      - Each FINAL SINGLE POINT ENERGY line (with surrounding separator lines)
-      - Each CARTESIAN GRADIENT block
-    """
-    lines = src.read_text().splitlines()
-    result: list[str] = []
-    charge_done = False
-    mult_done = False
-    i = 0
-
-    while i < len(lines):
-        line = lines[i]
-
-        # Charge line (emit once)
-        if not charge_done and "Total Charge" in line and "Charge" in line and "...." in line:
-            result.append(line)
-            charge_done = True
-            i += 1
-            continue
-
-        # Mult line (emit once)
-        if not mult_done and "Multiplicity" in line and "Mult" in line and "...." in line:
-            result.append(line)
-            mult_done = True
-            i += 1
-            continue
-
-        # CARTESIAN COORDINATES block
-        if "CARTESIAN COORDINATES (ANGSTROEM)" in line:
-            result.append("")
-            if i > 0 and _is_separator(lines[i - 1]):
-                result.append(lines[i - 1])
-            result.append(line)
-            i += 1
-            if i < len(lines):
-                result.append(lines[i])  # separator after header
-                i += 1
-            while i < len(lines) and lines[i].strip():
-                result.append(lines[i])
-                i += 1
-            result.append("")
-            continue
-
-        # FINAL SINGLE POINT ENERGY line (with surrounding separators)
-        if "FINAL SINGLE POINT ENERGY" in line:
-            result.append("")
-            if i > 0 and _is_separator(lines[i - 1]):
-                result.append(lines[i - 1])
-            result.append(line)
-            if i + 1 < len(lines) and _is_separator(lines[i + 1]):
-                result.append(lines[i + 1])
-                i += 2
-            else:
-                i += 1
-            result.append("")
-            continue
-
-        # CARTESIAN GRADIENT block (match exact header, not "CARTESIAN GRADIENT NORMS")
-        if line.strip() == "CARTESIAN GRADIENT":
-            result.append("")
-            if i > 0 and _is_separator(lines[i - 1]):
-                result.append(lines[i - 1])
-            result.append(line)
-            i += 1
-            if i < len(lines):
-                result.append(lines[i])  # separator after header
-                i += 1
-            if i < len(lines):
-                result.append(lines[i])  # blank line between separator and data
-                i += 1
-            while i < len(lines) and lines[i].strip():
-                result.append(lines[i])
-                i += 1
-            result.append("")
-            continue
-
-        i += 1
-
-    # Collapse consecutive blank lines into one
-    cleaned: list[str] = []
-    prev_blank = False
-    for ln in result:
-        if not ln.strip():
-            if not prev_blank:
-                cleaned.append(ln)
-            prev_blank = True
-        else:
-            cleaned.append(ln)
-            prev_blank = False
-
-    return "\n".join(cleaned).rstrip() + "\n"
-
-
-@dataclass(frozen=True)
-class OutFileExporter:
-    out_file: Path
-    enabled: bool
-
-    def export_from(self, src: Path) -> None:
-        """Extract grepper-relevant blocks from src and write to self.out_file."""
-        if not self.enabled:
-            return
-        content = _extract_orca_blocks(src)
-        self.out_file.write_text(content)
 
 
 @pytest.fixture(scope="session")
